+++ RS.Fritz.Manager.API/Services/TR-064/Services/WanCommonInterfaceConfig/Entities/WanCommonInterfaceConfigGetTotalBytesSentResponse.cs
@@ -2,4 +2,4 @@
 
 [MessageContract(WrapperName = "GetTotalBytesSentResponse")]
 public readonly record struct WanCommonInterfaceConfigGetTotalBytesSentResponse(
-    [property: MessageBodyMember(Name = "NewTotalBytesSent")] uint TotalBytesSent);
+    [property: MessageBodyMember(Name = "NewTotalBytesSent")] long TotalBytesSent);
